Order by date and then order by ID
When Instagram galleries are imported each post has the same exact publish timestamp. When paginating posts we get some funky behavior as the order of gallery posts is random from one page to the next. Here we modify the query so if the publish dates are all the same we fall back to sorting by post ID in descending order to ensure a consistent order from page to page.